fix: refactor global components
