+++ src/app/(frontend)/layout.tsx
@@ -4,14 +4,14 @@ import React from 'react'
 import type { Metadata } from 'next'
 import { draftMode } from 'next/headers'
 import { AdminBar } from '@/components/AdminBar'
-import { RenderFooter } from '@/Footer/RenderFooter'
+import { RenderFooter } from '@/globals/Footer/RenderFooter'
 import { Providers } from '@/providers'
 import { InitTheme } from '@/providers/Theme/InitTheme'
 import { mergeOpenGraph } from '@/utilities/mergeOpenGraph'
 import { cn } from '@/utilities/ui'
 import './globals.css'
-import { CustomCode } from '@/CustomCode/Component'
-import { RenderHeader } from '@/Header/RenderHeader'
+import { CustomCode } from '@/globals/CustomCode/Component'
+import { RenderHeader } from '@/globals/Header/RenderHeader'
 import { getServerSideURL } from '@/utilities/getURL'
 
 export default async function RootLayout({ children }: { children: React.ReactNode }) {
