+++ src/globals/Footer/components/Footer1/Component.tsx
