+++ src/globals/Footer/components/Footer1/config.ts
