+++ src/globals/Footer/components/Footer10/Component.tsx
