+++ src/globals/Footer/components/Footer10/config.tsx
