+++ src/globals/Footer/components/Footer2/Component.tsx
