+++ src/globals/Footer/components/Footer2/config.ts
