+++ src/globals/Footer/components/Footer3/Component.tsx
