+++ src/globals/Footer/components/Footer3/config.ts
