+++ src/globals/Footer/components/Footer4/config.ts
