+++ src/globals/Footer/components/Footer5/config.ts
