+++ src/globals/Footer/components/Footer6/Component.tsx
