+++ src/globals/Footer/components/Footer6/config.ts
