+++ src/globals/Footer/components/Footer7/Component.tsx
