+++ src/globals/Footer/components/Footer7/config.tsx
