+++ src/globals/Footer/components/Footer8/Component.tsx
