+++ src/globals/Footer/components/Footer8/config.tsx
