+++ src/globals/Footer/components/Footer9/config.tsx
