+++ src/globals/Footer/components/shared/FOOTER_DEVELOPMENT_GUIDE.md
