+++ src/globals/Footer/components/shared/ThemeEffect.tsx
