+++ src/globals/Footer/config.ts
