+++ src/globals/Footer/hooks/revalidateFooter.ts
