+++ src/globals/Footer/index.ts
