+++ src/globals/Header/components/Header1/config.ts
