+++ src/payload.config.ts
@@ -15,9 +15,9 @@ import { Pages } from './collections/Pages'
 import { Posts } from './collections/Posts'
 import { Tenants } from './collections/Tenants'
 import Users from './collections/Users'
-import { Footer } from './Footer/config'
 import { ApiKey } from './globals/api-keys'
-import { Header } from './Header/config'
+import { Footer } from './globals/Footer/config'
+import { Header } from './globals/Header/config'
 import { plugins } from './plugins'
 import { getServerSideURL } from './utilities/getURL'
